Prevent any move that would leave the player in check

+++ app/models/piece.rb
@@ -16,28 +16,52 @@ class Piece < ActiveRecord::Base
   def move_to!(new_x, new_y)
     return false unless game.current_player == color
     return false unless valid_move?(new_x, new_y)
-    destination_piece = game.piece_at(new_x, new_y)
 
     # If the destination piece is friendly, reject the move.
-    # Otherwise, capture the destination piece.
-    if destination_piece
+    # Otherwise, remove the destination piece for pending capture.
+    if (destination_piece = game.piece_at(new_x, new_y))
       return false if destination_piece.color == color
-      destination_piece.destroy
+      capturee_location = destination_piece_log(destination_piece)
+      return update_game_attributes(new_x, new_y, destination_piece, capturee_location)
     end
 
     update_game_attributes(new_x, new_y)
   end
 
-  def update_game_attributes(new_x, new_y)
+  def destination_piece_log(destination_piece)
+    result = [destination_piece.x_coord, destination_piece.y_coord]
+    destination_piece.update_attributes!(x_coord: 500, y_coord: 100)
+    result
+  end
+
+  def update_game_attributes(new_x, new_y, destination_piece = nil, capturee_location = nil)
+    original_status = [x_coord, y_coord, moved]
+
+    update_attributes!(x_coord: new_x, y_coord: new_y, moved: true)
+
+    if game.in_check?(color)
+      # Move is illegal: put everything back where it was.
+      reset_pieces!(original_status, destination_piece, capturee_location)
+      return false
+    end
+
+    destination_piece && destination_piece.destroy
+
     # If the move is not being made by a pawn, en passant capture is not
     # possible on the next move. If it is being made by a pawn, the move_to!
     # method in the Pawn class will set this value appropriately.
     game.update_attribute(:en_passant_file, nil) unless is_a? Pawn
 
-    update_attributes!(x_coord: new_x, y_coord: new_y, moved: true)
-
     # Assign turn to the other player after a successful move.
     game.switch_players!
+  end
+
+  def reset_pieces!(original_status, destination_piece, capturee_location)
+    update_attributes!(x_coord: original_status[0],
+                       y_coord: original_status[1],
+                       moved: original_status[2])
+    destination_piece && destination_piece.update_attributes!(x_coord: capturee_location[0],
+                                                              y_coord: capturee_location[1])
   end
 
   def valid_move?(new_x, new_y)
@@ -83,7 +107,7 @@ class Piece < ActiveRecord::Base
     x_offset == 0 || y_offset == 0 || diagonal_move?(new_x, new_y)
   end
 
-  # Returns the potenial move's displacement along each axis.
+  # Returns the potential move's displacement along each axis.
   def movement_by_axis(new_x, new_y)
     [(new_x - x_coord), (new_y - y_coord)]
   end
